+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for .NET Changelog
 
-## <a name="3.3.15-3.3.21"></a> 3.3.15-3.3.21 (2021-02-23)
+## <a name="3.3.15-3.3.22"></a> 3.3.15-3.3.22 (2021-04-05)
 
 ### Fixes
+* **clients** Fixed counter's method names
 * **clients** Fixed processing of empty routes
 * **clients** Fixed adding null filter/paging parameters